% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeroinfl.R
\name{dataClasses.zeroinfl}
\alias{dataClasses.zeroinfl}
\title{\link[ecip]{dataClasses} for \link[pscl]{zeroinfl} Object}
\usage{
\method{dataClasses}{zeroinfl}(x)
}
\arguments{
\item{x}{a \link[pscl]{zeroinfl} object}
}
\description{
\link[ecip]{dataClasses} for \link[pscl]{zeroinfl} Object
}
\note{
Function \code{pscl:::terms.zeroinfl()} return is not desired (not sure to which version tzh was commenting..).
}
\keyword{internal}
