% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeroinfl.R
\name{md_.zeroinfl}
\alias{md_.zeroinfl}
\title{R Markdown Lines for \link[pscl]{zeroinfl} Object}
\usage{
\method{md_}{zeroinfl}(x, xnm, ...)
}
\arguments{
\item{x, xnm, ...}{..}
}
\description{
R Markdown Lines for \link[pscl]{zeroinfl} Object
}
\examples{
library(ecip); list(
 '`zeroinfl`' = zeroinfl(art ~ . | 1, data = bioChemists)
) |> rmd.tzh::render_(file = 'zeroinfl')
}
\keyword{internal}
