% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibs.R
\name{pscl_bib}
\alias{pscl_bib}
\alias{.lambert92}
\title{bibs of \CRANpkg{pscl}}
\usage{
.lambert92(key = "Lambert92", ...)
}
\arguments{
\item{key, ...}{\link[utils]{bibentry}}
}
\description{
bibs of \CRANpkg{pscl}
}
\keyword{internal}
