% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeroinfl.R
\name{S3_zeroinfl}
\alias{S3_zeroinfl}
\alias{coef_.zeroinfl}
\alias{endpoint.zeroinfl}
\alias{desc_.zeroinfl}
\alias{.pval.summary.zeroinfl}
\alias{confint_.zeroinfl}
\title{Additional S3 methods for \link[pscl]{zeroinfl}}
\usage{
\method{coef_}{zeroinfl}(x)

\method{endpoint}{zeroinfl}(x)

\method{desc_}{zeroinfl}(x)

\method{.pval}{summary.zeroinfl}(x)

\method{confint_}{zeroinfl}(x, level = 0.95, ...)
}
\arguments{
\item{x}{\link[pscl]{zeroinfl}}

\item{level}{..}

\item{...}{..}
}
\description{
Additional S3 methods for \link[pscl]{zeroinfl}
}
\keyword{internal}
