% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeroinfl.R
\name{zeroinfl}
\alias{zeroinfl}
\title{\link[pscl]{zeroinfl} Objects}
\description{
\link[pscl]{zeroinfl} Objects
}
\examples{
library(ecip); list(
 '`zeroinfl`' = zeroinfl(art ~ . | 1, data = bioChemists)
) |> fastmd::render2html()

}
